refactor: Rename styled_stream module
Reflect that the primary value is not StyledStr
