+++ crates/anstyle-roff/src/styled_str.rs
